+++ lib/mason_from.dart
@@ -1,5 +1,0 @@
-/// To get a brickbrick from bundlebundle
-library mason_from;
-
-export 'src/commands/generator_no_substitution.dart' show CreateCommand;
-export 'src/mason_from.dart';