+++ lib/src/mason_from.dart
@@ -1,7 +1,0 @@
-/// {@template mason_from}
-/// A Very Good Project created by Very Good CLI.
-/// {@endtemplate}
-class MasonFrom {
-  /// {@macro mason_from}
-  const MasonFrom();
-}